+++ pkger-cli/src/opts.rs
@@ -50,6 +50,9 @@ pub enum Command {
         #[clap(short, long)]
         /// Disable colored output.
         raw: bool,
+        #[clap(short, long)]
+        /// Should the output be more verbose and include fields like version, arch...
+        verbose: bool,
     },
     /// Deletes the cache files with image state.
     CleanCache,
@@ -110,16 +113,8 @@ pub enum EditObject {
 
 #[derive(Debug, Subcommand)]
 pub enum ListObject {
-    Images {
-        #[clap(short, long)]
-        /// Should the output be more verbose and include additional fields.
-        verbose: bool,
-    },
-    Recipes {
-        #[clap(short, long)]
-        /// Should the output be more verbose and include fields like version, arch...
-        verbose: bool,
-    },
+    Images,
+    Recipes,
     Packages {
         #[clap(short, long)]
         images: Option<Vec<String>>,