+++ pkger-core/src/recipe/mod.rs
@@ -5,8 +5,9 @@ mod metadata;
 pub use cmd::Command;
 pub use envs::Env;
 pub use metadata::{
-    deserialize_images, BuildTarget, DebInfo, DebRep, Dependencies, Distro, GitSource, ImageTarget,
-    Metadata, MetadataRep, Os, PackageManager, Patch, Patches, PkgInfo, PkgRep, RpmInfo, RpmRep,
+    deserialize_images, BuildArch, BuildTarget, DebInfo, DebRep, Dependencies, Distro, GitSource,
+    ImageTarget, Metadata, MetadataRep, Os, PackageManager, Patch, Patches, PkgInfo, PkgRep,
+    RpmInfo, RpmRep,
 };
 
 use crate::{err, Error, Result};